+++ fuc/cli/vcf_vep.py
@@ -4,7 +4,7 @@ from argparse import RawTextHelpFormatter
 command = api.common.script_name(__file__)
 
 description = f"""
-This command will filter a VEP-annotated VCF file. It essentially wraps the `pandas.DataFrame.query` method. For details on query expression, please visit the method's documentation page (https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.query.html#pandas-dataframe-query).
+This command will filter a VCF file annotated by Ensemble VEP. It essentially wraps the `pandas.DataFrame.query` method. For details on query expression, please visit the method's documentation page (https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.query.html#pandas-dataframe-query).
 
 examples:
   $ fuc {command} in.vcf 'SYMBOL == "TP53"' > out.vcf
@@ -19,11 +19,11 @@ examples:
 def create_parser(subparsers):
     parser = subparsers.add_parser(
         command,
-        help='[VEP] filter a VEP-annotated VCF file',
+        help='[VCF] filter a VCF file annotated by Ensemble VEP',
         description=description,
         formatter_class=RawTextHelpFormatter,
     )
-    parser.add_argument('vcf', help='VEP-annotated VCF file')
+    parser.add_argument('vcf', help='Ensemble VEP-annotated VCF file')
     parser.add_argument('expr', help='query expression to evaluate')
     parser.add_argument(
         '--opposite',